Crate flight

Modules

draw
Mesh drawing

load
Asset loading

mesh
Mesh specification and upload

vr
VR hardware interface

Structs

Error
Combines all errors that can occur in the flight API

Light
Parameters for a point light source

Sun
Parameters for a sun light source

Texture
GPU-allocated texture object. Since this is just a reference to assets stored on the GPU, its memory footprint is negligible and it can be cloned freely.

Enums

ErrorKind
Error type enum

Traits

NativeRepr
Zero-cost conversion between a complex wrapper and its native form. For example, between &mut Vector2<f32> and &mut [f32; 2].

Type Definitions

ColorFormat
The pixel format of color drawing targets

DepthFormat
The pixel format of depth drawing targets

DepthRef
Reference to a GPU depth target

PbrMesh
A mesh that can be physically (realistically) rendered

ShaderResult
The result of compiling and linking shader programs

ShadowDepthFormat
The pixel format of shadow depth buffers

TargetRef
Reference to a GPU color target

UberMesh