Module flight::draw

Mesh drawing

Structs

DrawParams
Parameters to the draw system

EyeParams
Parameters that control the rendering of an eye

Painter
The painter is responsible for drawing meshes. Painters are instantiated with an associated style which specifies the data required for drawing (vertex type, material params, configuration) and implements the drawing pipeline. Note that a painter can only be used with primitive types that have been passed to setup.

PbrInputs
The configuration for physically based rendering

PbrMaterial
The collection of mesh textures used by physically based rendering

PbrStyle
Draws meshes using a physically based rendering pipeline

SolidInputs
The configuration for solid color rendering

SolidStyle
Draws objects in solid colors (without lighting) using the per-vertex color attribute

UberEnv
The scene environment

UberInputs
The configuration for physically based rendering

UberMaterial
The collection of mesh textures used by physically based rendering

UberStyle
Draws meshes using a physically based rendering pipeline

UnishadeInputs
The configuration for unishade rendering

UnishadeStyle
Draws objects with very simple lighting: one color on the top and a different color on the bottom

Constants

LIGHT_COUNT
The maximum number of point lights that can be simulated

Traits

Style
Implements a particular drawing process and visual style.

StyleInputs
Required configuration options for a Style