Struct flight::draw::Painter

pub struct Painter<R: Resources, E: Style<R>> { /* fields omitted */ }

The painter is responsible for drawing meshes. Painters are instantiated with an associated style which specifies the data required for drawing (vertex type, material params, configuration) and implements the drawing pipeline. Note that a painter can only be used with primitive types that have been passed to setup.

Methods

impl<R: Resources> Painter<R, UberStyle<R>>

pub fn clear_env<C: CommandBuffer<R>>(&self, ctx: &mut DrawParams<R, C>)

impl<R: Resources, E: Style<R>> Painter<R, E>

pub fn new<F: Factory<R> + FactoryExt<R>>(
f: &mut F
) -> Result<Painter<R, E>, Error>

Create a new painter in the given style and using the given factory.

pub fn setup<F: Factory<R> + FactoryExt<R>>(
&mut self,
f: &mut F,
prim: Primitive
) -> Result<(), Error>

Add the ability to draw the given primitive. This must be done before a mesh using the primitive is drawn.

pub fn try_draw<C>(
&self,
ctx: &mut DrawParams<R, C>,
model: Transform3<f32>,
mesh: &Mesh<R, E::Vertex, E::Material>
) -> Result<(), Error> where
C: CommandBuffer<R>,

Attempt to draw a mesh with the given parameters and model matrix, returning Err if something goes wrong.

pub fn draw<C>(
&self,
ctx: &mut DrawParams<R, C>,
model: Transform3<f32>,
mesh: &Mesh<R, E::Vertex, E::Material>
) where
C: CommandBuffer<R>,

Draw a mesh with the given parameters and model matrix, logging any errors.

pub fn cfg<F: FnOnce(&mut E::Inputs)>(&self, f: F)

Configure the draw style. For example, cfg(|c| c.ambient([1., 0., 0., 1.])) might set the ambient light color to red. The exact customization available depends on the style being used.