Struct flight::mesh::Mesh

pub struct Mesh<R: Resources, T: Vertex, M> {
    pub slice: Slice<R>,
    pub buf: Buffer<R, T>,
    pub prim: Primitive,
    pub mat: M,
}

A reference to a GPU mesh object that can be drawn by a Painter supporting the associated resource, vertex, material, and primitive combination. The memory cost of this object is negligible, since (unlike MeshSource) it is a cloneable reference to GPU allocated resources.

Fields
slice: Slice<R>

Reference to slice object (index buffer or range)

buf: Buffer<R, T>

Reference to VBO

prim: Primitive

Primitive type

mat: M

Material/texture data

Methods

impl<R: Resources, T: Vertex, M> Mesh<R, T, M>

pub fn with_material<N>(self, mat: N) -> Mesh<R, T, N>

Set the material of this mesh (usually just textures)