Struct flight::mesh::MeshSource

pub struct MeshSource<V, M> {
    pub verts: Vec<V>,
    pub inds: Indexing,
    pub prim: Primitive,
    pub mat: M,
}

A mesh storage object and builder that sent as drawable geometry to the GPU. Once uploaded, this mesh can only be rendered using a Painter whose associated style supports the given vertex and material type (<V, M>). In addition, the painter must be setup for this mesh's primitive type.

Fields
verts: Vec<V>

Vertices

inds: Indexing

Indexing scheme

prim: Primitive

Primitive type

mat: M

Material/texture data

Methods

impl<T: Vertex, M> MeshSource<T, M>

pub fn upload<R: Resources, F: FactoryExt<R>>(self, f: &mut F) -> Mesh<R, T, M>

Upload this mesh to the GPU.

pub fn with_material<N>(self, mat: N) -> MeshSource<T, N>

Set the material of this mesh (usually just textures)

impl<V: WithNorm, M> MeshSource<V, M>

pub fn with_normal(self, n: Vector3<f32>) -> MeshSource<V::With, M>

Adds the given normal vector to each vertex's attributes

impl<V: WithColor, M> MeshSource<V, M>

pub fn with_color(self, c: [f32; 3]) -> MeshSource<V::With, M>

Adds the given color to each vertex's attributes

impl<V: WithTex, M> MeshSource<V, M>

pub fn with_tex(self, c: Point2<f32>) -> MeshSource<V::With, M>

Adds the given texture or UV coordinates to each vertex's attributes

impl<V, M> MeshSource<V, M> where
V: WithTan + HasTex,
V::With: HasTex,

pub fn compute_tan(self) -> MeshSource<V::With, M>

Computes tangents and bitangents for a textured mesh so that normal mapping can be used. The calculated vectors will be 0 if the primitive type is not TriangleList or TriangleStrip.