Struct flight::Error

pub struct Error {
    pub context: Option<String>,
    pub error: ErrorKind,
}

Combines all errors that can occur in the flight API

Fields
context: Option<String> error: ErrorKind
Methods

impl Error

pub fn invalid_primitive(p: Primitive) -> Error

impl Error

pub fn cube_size(expected: u32) -> Error

impl Error

pub fn context(self, ctx: String) -> Error

Add some text explaining the context in which the error occurred