Struct flight::Texture

pub struct Texture<R, T> where
    R: Resources,
    T: TextureFormat,  {
    pub sampler: Sampler<R>,
    pub buffer: ShaderResourceView<R, <T as Formatted>::View>,
}

GPU-allocated texture object. Since this is just a reference to assets stored on the GPU, its memory footprint is negligible and it can be cloned freely.

Fields
sampler: Sampler<R> buffer: ShaderResourceView<R, <T as Formatted>::View>
Methods

impl<R: Resources, T: TextureFormat> Texture<R, T>

pub fn into_tuple(self) -> (ShaderResourceView<R, T::View>, Sampler<R>)

Convert this texture reference to an internally recognized tuple form

pub fn uniform_value<F>(
f: &mut F,
val: <<T as Formatted>::Surface as SurfaceTyped>::DataType
) -> Result<Self, Error> where
F: Factory<R>,

Build a single-pixel (single value) texture. The overhead on this might still be fairly high, even though the memory usage is minimal.