Trait flight::NativeRepr

pub trait NativeRepr<T: Sized + Copy>: Sized + Copy {
    fn upgrade(nat: T) -> Self { ... }

    fn upgrade_ref(nat: &T) -> &Self { ... }

    fn upgrade_mut(nat: &mut T) -> &mut Self { ... }

    fn downgrade(self) -> T { ... }

    fn downgrade_ref(&self) -> &T { ... }

    fn downgrade_mut(&mut self) -> &mut T { ... }
}

Zero-cost conversion between a complex wrapper and its native form. For example, between &mut Vector2<f32> and &mut [f32; 2].

Provided Methods

fn upgrade(nat: T) -> Self

Upgrade some native data into a high level wrapper

Important traits for &'a mut I
impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

fn upgrade_ref(nat: &T) -> &Self

Borrow some native data as a high level wrapper

Important traits for &'a mut I
impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

fn upgrade_mut(nat: &mut T) -> &mut Self

Mutably borrow some native data as a high level wrapper

fn downgrade(self) -> T

Downgrade this high level wrapper into some native data

Important traits for &'a mut I
impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

fn downgrade_ref(&self) -> &T

Borrow this high level wrapper as some native data

Important traits for &'a mut I
impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

fn downgrade_mut(&mut self) -> &mut T

Mutably borrow this high level wrapper as some native data