Module flight::vr

VR hardware interface

Structs

ControllerMoment
Instantaneous information about a controller. This can be used directly or to update some persistent state.

HmdMoment
Instantaneous information about the HMD. This can be used directly or to update some persistent state.

MappedController
A structure for tracking the state of a mapped controller.

VrContext
Provides access to VR hardware.

VrMoment
Instantaneous information about the VR system retrieved from VrContext::sync(). This can be used directly or to update some persistent state.

Enums

ControllerRef
Used to persistently identity a controller, either by internal id or by role. Note that roles can refer to different physical devices at different times, while the internal id will remain locked to a particular device.

Traits

Trackable
A device that provides instantaneous position and orientation information.

Functions

primary
Create a reference to the primary controller.

secondary
Create a reference to the secondary controller.

tertiary
Create a reference to the tertiary controller.

Type Definitions

ButtonMoment
Instantaneous information about a button.

ControllerIter
Iterator over momentary controller information.