Enum flight::vr::ControllerRef

pub enum ControllerRef {
    Primary,
    Secondary,
    Tertiary,
    Indexed(u32),
}

Used to persistently identity a controller, either by internal id or by role. Note that roles can refer to different physical devices at different times, while the internal id will remain locked to a particular device.

Variants
PrimarySecondaryTertiaryIndexed(u32)
Methods

impl ControllerRef

pub fn fixed(&self, moment: &VrMoment) -> ControllerRef

Make thus reference specific to a device (internal id) rather than dynamically updating (role).