Struct flight::vr::ControllerMoment

pub struct ControllerMoment {
    pub name: String,
    pub pose: Isometry3<f32>,
    pub axes: Vec<f64>,
    pub buttons: Vec<ButtonMoment>,
    // some fields omitted
}

Instantaneous information about a controller. This can be used directly or to update some persistent state.

Fields
name: String

The textual name of the controller

pose: Isometry3<f32>

The location and orientation of the controller

axes: Vec<f64>

The state of the floating point inputs on the controller

buttons: Vec<ButtonMoment>

The state of the button inputs on the controller

Methods

impl ControllerMoment

pub fn reference(&self) -> ControllerRef

Create a reference to this particular hardware device (not to its role).