Struct flight::vr::MappedController

pub struct MappedController {
    pub is: ControllerRef,
    pub last_timestamp: Option<f64>,
    pub dt: f64,
    pub connected: bool,
    pub pose: Isometry3<f32>,
    pub lin_vel: Vector3<f32>,
    pub ang_vel: Vector3<f32>,
    pub pose_delta: Isometry3<f32>,
    pub trigger: f64,
    pub trigger_delta: f64,
    pub pad: Point2<f64>,
    pub pad_delta: Vector2<f64>,
    pub pad_touched: bool,
    pub menu: bool,
    pub grip: bool,
    pub max_step: f64,
}

A structure for tracking the state of a mapped controller.

Fields
is: ControllerRef

The controller that updates this state object

last_timestamp: Option<f64>

Time of last update

dt: f64

Time since last update in seconds

connected: bool

The controller connection status.

pose: Isometry3<f32>

The pose of the controller

lin_vel: Vector3<f32>

The linear velocity of the controller

ang_vel: Vector3<f32>

The rotational axis of the controller multiplied by the rotation velocity (rad/s)

pose_delta: Isometry3<f32>

The transformation of the controller between the second most and most recent updates

trigger: f64

How far is the trigger pulled

trigger_delta: f64

The change in the trigger between the second most and most recent updates

pad: Point2<f64>

The last touched location on the circular pad

pad_delta: Vector2<f64>

The change in touch location on the circular pad between the second most and most recent updates

pad_touched: bool

Is the circular pad touched

menu: bool

Is the menu button pressed

grip: bool

Are the grip buttons pressed

max_step: f64

Maximum time step for velocity calculations

Methods

impl MappedController

pub fn new(reference: ControllerRef) -> MappedController

Create a simple default state that will be updated with data from the given controller.

pub fn update(&mut self, mom: &VrMoment) -> Result<(), ()>

Update the controller state using the provided instantaneous information.

pub fn pad_theta(&self) -> f64

Get the radial location of the last circular pad touch.