Struct flight::vr::VrContext

pub struct VrContext {
    pub near: f64,
    pub far: f64,
    // some fields omitted
}

Provides access to VR hardware.

Fields
near: f64

Set the eyes' near clipping plane

far: f64

Set the eyes' far clipping plane

Methods

impl VrContext

pub fn init(vrsm: VRServiceManager) -> Option<VrContext>

pub fn new() -> Option<VrContext>

Connect to default hardware devices.

pub fn mock() -> Option<VrContext>

Connect to a mock HMD.

pub fn set_texture(&mut self, texture_id: u32)

Set the OpenGL texture id to display on the HMD.

pub fn start(&mut self)

Start drawing to the HMD.

pub fn stop(&mut self)

Stop drawing to the HMD.

pub fn retrieve_size(&mut self) -> (u32, u32)

Retrieve the HMD device from the hardware API.

pub fn sync(&mut self) -> VrMoment

Synchronize with the hardware, returning transient details about the VR system at the specific moment in time. This data can be used directly or to update state variables.