Struct flight::vr::VrMoment

pub struct VrMoment {
    pub stage: Similarity3<f32>,
    pub inverse_stage: Similarity3<f32>,
    pub exit: bool,
    pub paused: bool,
    pub new_controllers: Vec<ControllerRef>,
    pub timestamp: f64,
    // some fields omitted
}

Instantaneous information about the VR system retrieved from VrContext::sync(). This can be used directly or to update some persistent state.

Fields
stage: Similarity3<f32>

The stage transform (moves the origin to the center of the room)

inverse_stage: Similarity3<f32>

The inverse stage transform (moves the center of the room to the origin)

exit: bool

Has the VR system requested the application to exit

paused: bool

Has the VR system requested the application to pause movement (should still sync and submit frames)

new_controllers: Vec<ControllerRef>

References to controllers that have connected since the last sync

timestamp: f64

Relative time of this moment (seconds)

Methods

impl VrMoment

pub fn controller(&self, role: ControllerRef) -> Option<&ControllerMoment>

Get a controller by reference if such a controller is connected.

pub fn controllers<'a>(&'a self) -> ControllerIter<'a>

Iterate over all connected controllers.

pub fn hmd(&self) -> Option<&HmdMoment>

Get instantaneous information about the HMD if it is connected.

pub fn submit(self, ctx: &mut VrContext)

Submit the rendered scene. This ends the applicability of this information, since it only applies to the state of the VR system at the last sync.